refact: removes v2 and integrates changes in team production system
