+++ team_production_system/models/availability.py
@@ -1,6 +1,7 @@
 from .mentor import Mentor
 from django.db.models.constraints import UniqueConstraint
 from django.db import models
+from datetime import timedelta
 
 
 # Allow mentors to set their avaliabiltiy
@@ -22,3 +23,20 @@ class Availability(models.Model):
             f"{self.mentor} is available from "
             f"{self.start_time} to {self.end_time}."
         )
+
+    def save(self, *args, **kwargs):
+            all_availabilities = []
+
+            # Create a list of all availabilities in 30 minute increments
+            while self.start_time < self.end_time:
+                all_availabilities.append(
+                    Availability(
+                        mentor=self.mentor,
+                        start_time=self.start_time,
+                        end_time=self.start_time + timedelta(minutes=30)
+                    )
+                )
+                self.start_time += timedelta(minutes=30)
+            
+            # Bulk create all availabilities
+            Availability.objects.bulk_create(all_availabilities)
