--- v2/models/__init__.py
@@ -1,9 +1,0 @@
-# flake8: noqa
-
-
-from .availability import Availability
-from .mentee import Mentee
-from .mentor import Mentor
-from .notification_settings import NotificationSettings
-from .session import Session
-from .custom_user import CustomUser
