--- v2/serializers/__init__.py
@@ -1,10 +1,0 @@
-# flake8: noqa
-
-from .availability import AvailabilitySerializer
-from .custom_user import CustomUserSerializer
-from .mentor_profile import MentorProfileSerializer
-from .mentor_list import MentorListSerializer
-from .mentee_profile import MenteeProfileSerializer
-from .mentee_list import MenteeListSerializer
-from .session import SessionSerializer
-from .notification_settings import NotificationSettingsSerializer
