--- v2/serializers/custom_user.py
@@ -1,22 +1,0 @@
-from rest_framework import serializers
-from team_production_system.models import CustomUser
-
-
-# The serializer for the user information
-class CustomUserSerializer(team_production_system.serializers.CustomUserSerializer):
-    profile_photo = serializers.ImageField()
-
-    class Meta:
-        model = CustomUser
-        fields = [
-            'pk',
-            'username',
-            'first_name',
-            'last_name',
-            'email',
-            'phone_number',
-            'profile_photo',
-            'is_mentor',
-            'is_mentee',
-            'is_active',
-        ]
