--- v2/serializers/mentee_list.py
@@ -1,14 +1,0 @@
-from rest_framework import serializers
-from team_production_system.models import CustomUser
-from .mentee_profile import MenteeProfileSerializer
-from .custom_user import CustomUserSerializer
-
-
-class MenteeListSerializer(team_production_system.serializers.MenteeListSerializer):
-    mentee_profile = MenteeProfileSerializer(read_only=True, source='mentee')
-    user = CustomUserSerializer(read_only=True, source='customuser')
-
-    class Meta:
-        model = CustomUser
-        fields = ('user', 'pk', 'username', 'first_name',
-                  'last_name', 'is_mentee', 'mentee_profile')
